+++ src/adapters/outbound/formatters/mod.rs
@@ -5,8 +5,6 @@ mod diff_markdown_formatter;
 mod markdown_formatter;
 
 pub use cyclonedx_formatter::CycloneDxFormatter;
-#[allow(unused_imports)]
 pub use diff_json_formatter::DiffJsonFormatter;
-#[allow(unused_imports)]
 pub use diff_markdown_formatter::DiffMarkdownFormatter;
 pub use markdown_formatter::MarkdownFormatter;